Add modular password toggle utility
