+++ Areas/Admin/Pages/Users/Create.cshtml
@@ -20,10 +20,15 @@
       <label asp-for="Input.Password" class="form-label"></label>
       <div class="password-container">
         <input asp-for="Input.Password" class="form-control" autocomplete="new-password" />
-        <button type="button" class="password-toggle" aria-label="Show password">
-          <svg viewBox="0 0 16 16" fill="none" stroke="currentColor" xmlns="http://www.w3.org/2000/svg">
+        <button type="button" class="password-toggle" aria-label="Show password" aria-pressed="false">
+          <svg class="eye-open" viewBox="0 0 16 16" fill="none" stroke="currentColor" xmlns="http://www.w3.org/2000/svg">
             <path d="M1 8s3-5 7-5 7 5 7 5-3 5-7 5-7-5-7-5z"/>
             <circle cx="8" cy="8" r="3"/>
+          </svg>
+          <svg class="eye-closed" viewBox="0 0 16 16" fill="none" stroke="currentColor" xmlns="http://www.w3.org/2000/svg" hidden>
+            <path d="M1 8s3-5 7-5 7 5 7 5-3 5-7 5-7-5-7-5z"/>
+            <circle cx="8" cy="8" r="3"/>
+            <path d="M1 1l14 14"/>
           </svg>
         </button>
       </div>
